pyproject.toml
[tool.poetry]
name = "pymqtt-demo"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"
packages = [{include = "pymqtt_demo"}]

[tool.poetry.dependencies]
python = "^3.9"
paho-mqtt = "^1.6.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
